+++ nengo/base.py
@@ -1,13 +1,14 @@
+from copy import copy
 import sys
 import warnings
 
 import numpy as np
 
 from nengo.config import Config
-from nengo.exceptions import ValidationError
+from nengo.exceptions import NotAddedToNetworkWarning, ValidationError
 from nengo.params import (
-    Default, FrozenObject, is_param, IntParam, NumberParam, Parameter,
-    StringParam, Unconfigurable)
+    CopyableObject, Default, FrozenObject, is_param, IntParam, NumberParam,
+    Parameter, StringParam, Unconfigurable)
 from nengo.rc import rc
 from nengo.utils.compat import is_integer, range, reraise, with_metaclass
 from nengo.utils.numpy import as_shape, maxint
@@ -34,7 +35,7 @@ class NetworkMember(type):
         return inst
 
 
-class NengoObject(with_metaclass(NetworkMember)):
+class NengoObject(with_metaclass(NetworkMember, CopyableObject)):
     """A base class for Nengo objects.
 
     Parameters
@@ -56,14 +57,35 @@ class NengoObject(with_metaclass(NetworkMember)):
     seed = IntParam('seed', default=None, optional=True)
 
     def __init__(self, label, seed):
+        super(NengoObject, self).__init__()
         self.label = label
         self.seed = seed
 
     def __getstate__(self):
-        raise NotImplementedError("Nengo objects do not support pickling")
+        state = super(NengoObject, self).__getstate__()
+        del state['_initialized']
+        return state
 
     def __setstate__(self, state):
-        raise NotImplementedError("Nengo objects do not support pickling")
+        from nengo.network import Network
+        super(NengoObject, self).__setstate__(state)
+        setattr(self, '_initialized', True)
+        if len(Network.context) > 0:
+            warnings.warn(
+                "{obj} was not added to the network. When copying objects, "
+                "use the copy method on the object instead of Python's copy "
+                "module. When unpickling objects, they have to be added to "
+                "networks manually.".format(obj=self),
+                NotAddedToNetworkWarning)
+
+    def copy(self, add_to_container=True):
+        with warnings.catch_warnings():
+            warnings.simplefilter('ignore', category=NotAddedToNetworkWarning)
+            c = copy(self)
+        if add_to_container:
+            from nengo.network import Network
+            Network.add(c)
+        return c
 
     def __setattr__(self, name, val):
         if hasattr(self, '_initialized') and not hasattr(self, name):